+++ Cargo.toml
@@ -25,7 +25,7 @@ rustc-hash = "1.1.0"
 miette = { version = "5.5.0", features = ["fancy"] }
 thiserror = "1.0.38"
 
-kanata-keyberon = "0.10.0"
+kanata-keyberon = "0.11.0"
 # Uncomment below and comment out above for testing local keyberon changes
 # kanata-keyberon = { path = "keyberon" }
